Change fill algorithm to forward fill, place items in priority order, open up levels logically
Fill was failing so made changes to place important items first, and open levels in the DK Isles logic as the beginning is very restrictive. Subject to change.
Also changed to Forward Fill. It helped with failures but ended up not being necessary, however it is ~3 faster. Kept Assumed Fill function so it'll be easy to change back