Update pointer table allocation algorithm
Updates pointer table allocation algorithm so that after a pointer table is re-allocated, it'll free up the space in ROM that used to house that pointer table. This means that pointer tables can be placed into locations that other pointer tables once were. By reversing the re-allocation order, it means that it goes through the pointer tables in reverse order, meaning that if pointer table 25 (the biggest table) is being reallocated, that space will be freed up first, usually resulting in the smallest ROM as every other re-allocation can (usually) fit within the old pointer 25 space.