Plandomizer update to handle starting moves
Moves that are selected as starting moves will not be selectable as plandomized rewards. (For progressive moves, if the total of assigned rewards + starting moves exceeds the maximum, the seed will not generate.)

There are some other fixes as well. Settings strings no longer break if the plandomizer is being used.