Plandomizer: fix client-side validation for item counting
If any starting moves are in a pool where every move is guaranteed to be given, they will count against the item counts for plandomizer client-side validation. Users will be warned if they try to place them.

Note to self: figure out how to stop a seed from genning if it fails client-side validation.