We're doing it live
- Added debugging output to hopefully narrow down an issue where cosmetic/music settings do not save.

If you're having that particular issue:
Please check your console log and report its output on your first and second load. (copy paste the part of the output that spams the settings names, please)

This update does nothing else, and is for debugging purposes only. (it's also very much temporary)