Fix for music plando
The added "let" keyword meant that we were creating and updating a local variable, but we needed to update the global cosmetic_truncated_names object instead. With that object empty, we were unable to find custom songs that corresponded to the user's selections in the dropdowns, so no custom music could be assigned.