feat: option to avoid placing hints for item locations in shops
In standard settings, hints relating to shops, with the exception of
path hints for purchased items, are generally quite weak;
the types of items sold in each shop can be scouted quickly
and without any prerequisite items.
I reckon there are two main offenders in season 4 settings: path hints
for a single key in a shop, and kongs sold in a shop.
Foolish hints for shops are also weak.

This commit introduces a game setting that, when enabled, causes
these three classes of hints to be avoided.